refactor : JWT 부분 SWM-KAWAI-MANS:jwt-manager로 대체
