+++ src/test/java/online/partyrun/partyrunmatchingservice/global/security/AuthUserTest.java
@@ -3,8 +3,7 @@ package online.partyrun.partyrunmatchingservice.global.security;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 
-import online.partyrun.partyrunmatchingservice.global.security.jwt.JwtPayload;
-
+import online.partyrun.jwtmanager.dto.JwtPayload;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
